fix: CandidateRequestDto ballotId type Long으로 수정

+++ capstone-design/src/main/java/com/inha/capstonedesign/voting/dto/request/CandidateRequestDto.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotNull;
 public class CandidateRequestDto {
 
     @NotNull
-    private Integer ballotId;
+    private Long ballotId;
     @NotNull
     private String candidateName;
 }
